fix(github): observe origin repository probe failures
